refactor(persistence): keep the in-memory backend out of inline test builds
Two follow-ups from review of the InMemoryStore extraction.

kanban-persistence gained a dev-dependency on kanban-backend-memory, which
depends back on kanban-persistence through kanban-backend. Cargo permits
that cycle, but an inline cfg(test) module builds kanban-persistence twice
and the two instances are distinct types. The one test that needed a
concrete DataStore moves to tests/, where the crate is linked once.

kanban-service took kanban-backend-memory as an unconditional dependency,
but its only production users are json_backend.rs and sqlite_backend.rs,
both already behind the json and sqlite features. It is now optional and
enabled by those features, matching how the sibling persistence crates are
declared, so building without them yields a service crate with no concrete
backend dependencies at all.
