+++ crates/kanban-service/Cargo.toml
@@ -10,8 +10,8 @@ description = "Shared service layer implementing KanbanOperations over a pluggab
 
 [features]
 default = ["json", "sqlite"]
-json = ["dep:kanban-persistence-json"]
-sqlite = ["dep:kanban-persistence-sqlite"]
+json = ["dep:kanban-persistence-json", "dep:kanban-backend-memory"]
+sqlite = ["dep:kanban-persistence-sqlite", "dep:kanban-backend-memory"]
 test-helpers = ["kanban-persistence/test-helpers"]
 # Derive `schemars::JsonSchema` on the wire request DTOs so MCP tool handlers
 # can use them directly as `Parameters<T>` (rmcp requires a JSON Schema). Kept
@@ -24,7 +24,7 @@ kanban-domain = { path = "../kanban-domain", version = "^0.7" }
 kanban-persistence = { path = "../kanban-persistence", version = "^0.7" }
 kanban-api = { path = "../kanban-api", version = "^0.7" }
 kanban-backend = { path = "../kanban-backend", version = "^0.7" }
-kanban-backend-memory = { path = "../kanban-backend-memory", version = "^0.7" }
+kanban-backend-memory = { path = "../kanban-backend-memory", version = "^0.7", optional = true }
 kanban-persistence-json = { path = "../kanban-persistence-json", version = "^0.7", optional = true }
 kanban-persistence-sqlite = { path = "../kanban-persistence-sqlite", version = "^0.7", optional = true }
 async-trait.workspace = true
